Fix boot configuration files to build an empty stub on darwin.
